moved functions/ and tests/ within app/ directory so flask app could find our custom methods
