+++ app/tests/test_rle.py
@@ -1,9 +1,9 @@
 import os
 import sys
 
-sys.path.insert(0, os.path.abspath('./functions'))
+sys.path.insert(0, os.path.abspath('.'))
 
-from rle import *
+from app.functions.rle import *
 
 string = 'AABBCDDDE'
